actually cool menu now

+++ src/gui/ugui.h
@@ -5,7 +5,7 @@ typedef struct engine_s engine_t;
 
 void ugui_mainmenu_bar(engine_t *);
 void ugui_mainmenu_bookmark(engine_t *, float x);
-void ugui_mainmenu_icon(engine_t *, float x, const char *label, int icon, int font, int active);
+void ugui_mainmenu_icon(engine_t *, float x, const char *label, int icon, int font, float active);
 
 #endif
